+++ src/routes/api/v2/slayers.js
@@ -1,0 +1,77 @@
+import * as helper from "../../../helper.js";
+import * as lib from "../../../lib.js";
+import cors from "cors";
+import express from "express";
+
+import { db } from "../../../mongo.js";
+
+const router = express.Router();
+
+router.use((req, res, next) => {
+  req.options.debugId = helper.generateDebugId("api/v2/slayers");
+  next();
+});
+
+router.get("/:player/:profile", cors(), async (req, res, next) => {
+  try {
+    const { profile, allProfiles } = await lib.getProfile(db, req.params.player, null, req.options);
+
+    let output = {
+      error: "Invalid Profile Name!",
+    };
+
+    for (const singleProfile of allProfiles) {
+      const cute_name = singleProfile.cute_name;
+
+      if (cute_name.toLowerCase() != req.params.profile.toLowerCase()) {
+        continue;
+      }
+
+      const items = await lib.getItems(singleProfile.members[profile.uuid], false, "", req.options);
+      const data = await lib.getStats(db, singleProfile, allProfiles, items, req.options);
+
+      output = {
+        profile_id: singleProfile.profile_id,
+        cute_name: singleProfile.cute_name,
+        slayer_xp: data.slayer_xp,
+        slayers: data.slayers,
+        slayer_bonus: data.slayer_bonus,
+        slayer_coins_spent: data.slayer_coins_spent,
+      };
+    }
+
+    res.json(output);
+  } catch (e) {
+    next(e);
+  }
+});
+
+router.get("/:player", cors(), async (req, res, next) => {
+  try {
+    const { profile, allProfiles } = await lib.getProfile(db, req.params.player, null, req.options);
+
+    const output = { profiles: {} };
+
+    for (const singleProfile of allProfiles) {
+      const userProfile = singleProfile.members[profile.uuid];
+
+      const items = await lib.getItems(userProfile, false, "", req.options);
+      const data = await lib.getStats(db, singleProfile, allProfiles, items, req.options);
+
+      output.profiles[singleProfile.profile_id] = {
+        profile_id: singleProfile.profile_id,
+        cute_name: singleProfile.cute_name,
+        slayer_xp: data.slayer_xp,
+        slayers: data.slayers,
+        slayer_bonus: data.slayer_bonus,
+        slayer_coins_spent: data.slayer_coins_spent,
+      };
+    }
+
+    res.json(output);
+  } catch (e) {
+    next(e);
+  }
+});
+
+export { router };
